transfer item report

+++ saleor/countertransfer/api/views.py
@@ -37,7 +37,11 @@ class DestroyView(generics.DestroyAPIView):
         for item in items:
             Stock.objects.increase_stock(item.stock, item.qty)
         # raise serializers.ValidationError('You cannot delete ')
-        instance.delete()
+        if instance.any_closed():
+            instance.trashed = True
+            instance.save()
+        else:
+            instance.delete()
 
 
 class DestroyItemView(generics.DestroyAPIView):
@@ -46,7 +50,11 @@ class DestroyItemView(generics.DestroyAPIView):
     def perform_destroy(self, instance):
         Stock.objects.increase_stock(instance.stock, instance.qty)
         # raise serializers.ValidationError('You cannot delete ')
-        instance.delete()
+        if instance.closed:
+            instance.trashed = True
+            instance.save()
+        else:
+            instance.delete()
 
 
 class ListAPIView(generics.ListAPIView):
@@ -64,13 +72,12 @@ class ListAPIView(generics.ListAPIView):
         return {"date": None, 'request': self.request}
 
     def get_queryset(self, *args, **kwargs):
+        queryset_list = Table.objects.filter(trashed=False)
         try:
             if self.kwargs['pk']:
-                queryset_list = Table.objects.filter(customer__pk=self.kwargs['pk']).order_by('car').distinct('car').select_related()
-            else:
-                queryset_list = Table.objects.all.select_related()
+                queryset_list = queryset_list.filter(customer__pk=self.kwargs['pk']).order_by('car').distinct('car').select_related()
         except Exception as e:
-            queryset_list = Table.objects.all()
+            pass
 
         page_size = 'page_size'
         if self.request.GET.get(page_size):
@@ -114,13 +121,12 @@ class ListItemsAPIView(generics.ListAPIView):
         return {"date": None, 'request': self.request}
 
     def get_queryset(self, *args, **kwargs):
+        queryset_list = Item.objects.filter(trashed=False)
         try:
             if self.kwargs['pk']:
-                queryset_list = Item.objects.filter(transfer__pk=self.kwargs['pk']).select_related()
-            else:
-                queryset_list = Item.objects.all.select_related()
+                queryset_list = queryset_list.filter(transfer__pk=self.kwargs['pk']).select_related()
         except Exception as e:
-            queryset_list = Item.objects.all()
+            pass
 
         page_size = 'page_size'
         if self.request.GET.get(page_size):
